requirements.txt
requests>=2.28.1
httpx>=0.23.0
beautifulsoup4>=4.11.1
lxml>=4.9.1
colorama>=0.4.5
tqdm>=4.64.0
tabulate>=0.8.10
cryptography>=38.0.1
aiohttp>=3.8.3
rich>=12.6.0
click>=8.1.3
python-dotenv>=0.21.0
pydantic>=1.10.2
websockets>=10.4
pymysql>=1.0.2
psycopg2-binary>=2.9.5
pyodbc>=4.0.35
oracledb>=1.3.1
pyyaml>=6.0
pandas>=1.5.3
numpy>=1.24.2
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.13.1
dash>=2.9.1
redis>=4.5.1
elasticsearch>=8.6.2
pymongo>=4.3.3
sqlalchemy>=2.0.7
aiokafka>=0.8.0
confluent-kafka>=2.0.2
grpcio>=1.53.0
protobuf>=4.22.1
graphene>=3.2.1
fastapi>=0.95.0
uvicorn>=0.21.1